requirements.txt
pymongo==3.10.1
Scrapy==1.5.1
lxml==4.6.2
jieba
wordcloud
pillow
matplotlib
snownlp
scikit-learn